Convert data from all samples in an Analysis

The function collects all samples data from an Analysis using the collect_one_sample_data() function and save them on disk.

Usage

convert_all_samples_data(
  analysis_id,
  connection_params = NULL,
  format = "parquet",
  path = NULL,
  overwrite = T,
  num_spectras = NULL
)

Arguments

analysis_id

The id of the analysis

connection_params

OPTIONAL: Connection parameters object created by the create_connection_params function. If not provided, the get_connection_params will look for such object in the global environment

format

The format chosen for the exported file (Parquet or HDF5)

path

OPTIONAL The destination path for exported files

overwrite

OPTIONAL overwrite the sample if already present on disk

num_spectras

OPTIONAL A number of spectra to be collected and converted (if no need to convert whole samples, e.g. for testing purposes)

Value

A dataframe for each sample is saved in Parquet or HDF5 format in a folder named after the Analysis.