Class containing a sample data and metadata

Contains sample data (as data.table in RAM or as pointer to Parquet file), metadata and spectrum metadata in table and json formats.

Usage

# S4 method for class 'sample_dataset'
get_sample_data(obj)

Arguments

obj

The sample_dataset object to access.

Value

get_sample_data returns a data.table object containing the sample data.

Details

Objects for this class are returned by collect_one_sample_data and create_sample_dataset.

Methods (by generic)

get_sample_data(sample_dataset): Accessor method to obtain the sample_data table.

Slots

sample_data

Contains a datatable with the sample data, or an Arrow Dataset R6 object pointing to the Parquet file data.

sample_metadata

Contains a datatable with the sample metadata.

spectrum_metadata

Contains a datatable with the spectrum metadata.

sample_metadata_json

Contains a character with the sample metadata.

spectrum_metadata_json

Contains a character with the spectrum metadata.

Use the collect_one_sample_data to

store the sample data, metadata and spectrum metadata, in table and json formats.