Class containing a sample's information

Contains sample metadata and spectrum metadata in table and json formats.

Usage

# S4 method for class 'sample_infos'
get_sample_metadata(obj)

# S4 method for class 'sample_infos'
get_sample_name(obj)

# S4 method for class 'sample_infos'
get_sample_id(obj)

# S4 method for class 'sample_infos'
get_analysis_name(obj)

# S4 method for class 'sample_infos'
get_sample_metadata_json(obj)

# S4 method for class 'sample_infos'
get_spectrum_metadata(obj)

# S4 method for class 'sample_infos'
get_spectrum_metadata_json(obj)

# S4 method for class 'sample_infos'
add_sample_id(obj, id)

Arguments

obj

The sample_infos object to access.

id

The identifier of the sample result for which spectrum information is to be retrieved.

Value

get_sample_metadata returns a data.table object containing the sample metadata.

get_sample_name returns a character object containing the sample name.

get_sample_id returns a character object containing the sample id.

get_analysis_name returns a character object containing the analysis name.

get_sample_metadata_json returns a json character object containing the sample metadata.

get_spectrum_metadata returns a data.table object containing the spectrum metadata of a sample.

get_spectrum_metadata_json returns a json character object containing the spectrum metadata of a sample.

add_sample_id returns a character object containing the sample id.

Details

Objects for this class are returned by get_sample_infos.

Methods (by generic)

get_sample_metadata(sample_infos): Accessor method to obtain the sample_infos table.

get_sample_name(sample_infos): Accessor method to obtain the sample name.

get_sample_id(sample_infos): Accessor method to obtain the sample id.

get_analysis_name(sample_infos): Accessor method to obtain the analysis name.

get_sample_metadata_json(sample_infos): Accessor method to obtain the sample metadata in json format.

get_spectrum_metadata(sample_infos): Accessor method to obtain the spectrum_metadata table.

get_spectrum_metadata_json(sample_infos): Accessor method to obtain the spectrum metadata in json format.

add_sample_id(sample_infos): Method to add a sample ID to the sample_infos object.

Slots

sample_metadata

Contains a datatable with the sample metadata.

spectrum_metadata

Contains a datatable with the spectrum metadata.

sample_metadata_json

Contains a character with the sample metadata.

spectrum_metadata_json

Contains a character with the spectrum metadata.

Use the get_sample_infos to

store the sample metadata and spectrum metadata, in table and json formats.