Save collected data from one sample

The function saves data collected from one sample with the collect_one_sample_data function. Datatables of spectra and metadata are saved in either two Parquet files or one HDF5 file.

Usage

save_one_sample_data(
  sample_dataset,
  sample_name = NULL,
  analysis_name = NULL,
  path = NULL,
  format = "parquet"
)

Arguments

sample_dataset

Collected data from a sample, stored in a sample_dataset object

sample_name

OPTIONAL The sample name (to be used for naming the saved file)

analysis_name

OPTIONAL The name of the Analysis (to be used for naming the folder)

path

OPTIONAL The destination path for the exported file

format

The format chosen for the exported file (parquet or hdf5)

Value

Datatables of the sample's spectral data and metadata are saved in Parquet or HDF5 format in the analysis name folder.

See also